+++ src/components/wishlist/index.ts
@@ -1,0 +1,5 @@
+export { default as WishlistEmpty } from "./WishlistEmpty";
+export { default as WishlistGrid } from "./WishlistGrid";
+export { default as WishlistHeader } from "./WishlistHeader";
+export { default as WishlistItem } from "./WishlistItem";
+export { default as WishlistSkeleton } from "./WishlistSkeleton";
